Implement poly and sympoly parsing
This is still a bit rough, but the basic principle is working. This
parser uses the same implementations for both poly and sympoly by
swapping the variable reader function. The parsing is not perfect
against wrong inputs yet, but this is enough to do some testing.

Renamed the two example files after the braid3_test.ms file, they all
form the same problem now.