Change some unordered_flat_map to unordered_map
This was a performance regression only noticable in debug mode that is
worse when using Clang. I expect this to stem from the flat map having
a different access pattern.

Interestingly, the flat map in the monomial and polynomial store is
really important, otherwise we take a bigger hit if we change it to a
regular unordered_map.